Ignore Stan estimators because GCC installation fails.
