+++ pymare/tests/test_estimators.py
@@ -2,8 +2,7 @@ import numpy as np
 
 from pymare.estimators import (WeightedLeastSquares, DerSimonianLaird,
                                VarianceBasedLikelihoodEstimator,
-                               SampleSizeBasedLikelihoodEstimator,
-                               StanMetaRegression)
+                               SampleSizeBasedLikelihoodEstimator)
 
 
 def test_weighted_least_squares_estimator(dataset):
@@ -64,14 +63,3 @@ def test_sample_size_based_restricted_maximum_likelihood_estimator(dataset_n):
     assert np.allclose(beta, [-2.1071], atol=1e-4)
     assert np.allclose(sigma2, 13.048, atol=1e-4)
     assert np.allclose(tau2, 3.2177, atol=1e-4)
-
-
-def test_stan_estimator(dataset):
-    # no ground truth here, so we use sanity checks and rough bounnds
-    results = StanMetaRegression(iter=2500).fit(dataset)
-    assert 'BayesianMetaRegressionResults' == results.__class__.__name__
-    summary = results.summary(['beta', 'tau2'])
-    beta1, beta2, tau2 = summary['mean'].values[:3]
-    assert -0.5 < beta1 < 0.1
-    assert 0.6 < beta2 < 0.9
-    assert 2 < tau2 < 6
